# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Telco Digital Credit Platform API
  description: |
    Platform-facing API. Tenant context is resolved from the authenticated
    credential (V2-TEN-002), never from caller-supplied fields; a conflicting
    `X-Telco-Id` claim is rejected 403 and security-audited (V2-TEN-003, EDG-026).

    Conventions (binding, V2-API-002..011):
    - Monetary values are integer **minor units** plus ISO 4217 currency; never floats.
    - Mutating commands require an `Idempotency-Key`; a valid retry returns the
      original outcome (V2-API-003).
    - Timestamps are UTC RFC 3339.
    - Errors use the standard envelope with stable `error_code` families
      (AUTH_*, TENANT_*, OFFER_*, ADVANCE_*, FULFILMENT_*, FUNDING_*, RATE_*, SYSTEM_*).
  version: 0.4.6  # M4d — finance settlement verify (scope-bounded statements + reproducibility)
servers:
  - url: http://localhost:8090
    description: Local development (A-14)
security:
  - ApiKeyAuth: []
paths:
...
  /v1/offers:
    post:
      operationId: listOffers
      summary: Current valid offers for a subscriber (V2-OFR-009 reuse)
      description: >
        Every value derives from governed product config + the pinned decision
        snapshot. Generation is serialized per subscriber/programme — repeated
        or concurrent enquiries return the same ladder. BX-MED-007: this read is a
        POST because it is keyed by the subscriber msisdn_token, a sensitive
        identifier that must never travel in a URL (query strings leak into browser
        history, access logs, proxies and telemetry).
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [programme_id, msisdn_token]
              properties:
                programme_id: { type: string }
                msisdn_token: { type: string }
      responses:
        "200":
          description: Offer ladder (may be empty when nothing is offerable)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Offer" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Unknown subscriber token (OFFER_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/advances:
    post:
      operationId: confirmAdvance
      summary: Confirm an offer — the origination saga (V2 §13)
      description: >
        Requires Idempotency-Key (V2-API-002): a retry with the same key
        returns the ORIGINAL outcome (200), never a second advance. 201 =
        credited (ACTIVE). 202 = PROCESSING — the outcome is being resolved
        (never blind-retried); poll status_route (EDG-004/005). 422 = telco
        definitively rejected. Customer-safe statuses only (V2-ADV-016).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [programme_id, offer_id, msisdn_token, disclosure_ref, channel, session_id, accepted_at]
              properties:
                programme_id: { type: string }
                offer_id: { type: string }
                msisdn_token: { type: string }
                disclosure_ref:
                  type: string
                  description: >
                    R-P0-7: the disclosure_ref from the offer, echoed back as
                    proof of the exact terms the customer was shown. Must match
                    the offer's disclosure snapshot or the confirm is refused.
                channel:
                  type: string
                  description: The channel the disclosure was accepted through; must be one the programme's disclosure policy permits (de-hardcodes USSD).
                session_id:
                  type: string
                  description: Telco-supplied channel session identifier (DD-06).
                accepted_at:
                  type: string
                  format: date-time
                  description: When the customer accepted; must fall within the disclosure's validity window.
                telco_evidence:
                  type: object
                  nullable: true
                  description: Optional telco-supplied acceptance signature/evidence (DD-06).
      responses:
        "201":
          description: Advance ACTIVE — value credited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Advance" }
        "200":
          description: Idempotent replay of the original outcome (EDG-001)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Advance" }
        "202":
          description: PROCESSING — resolution in progress, poll status_route
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Advance" }
        "400":
          description: Missing/invalid Idempotency-Key or malformed body
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Subscriber not eligible (SUBSCRIBER_INACTIVE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Offer not found (OFFER_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: >
            OFFER_EXPIRED | OFFER_SNAPSHOT_MISMATCH |
            CONCURRENT_ADVANCE_BLOCK | FUNDING_POOL_EXHAUSTED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Telco definitively rejected fulfilment (status FAILED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Advance" }
  /v1/advances/{id}:
    get:
      operationId: getAdvanceStatus
      summary: Durable advance status (EDG-004 — survives dropped sessions)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Customer-safe advance status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Advance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Unknown advance (ADVANCE_NOT_FOUND; tenant-scoped visibility)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/self-exclusions:
    post:
      operationId: requestSelfExclusion
      summary: Self-exclude a subscriber from credit (R1-MUST, responsible lending)
      description: >
        Records a subscriber's opt-out from being offered/receiving credit and
        enforces it at the offer/confirm gate. Idempotent — a repeat request
        returns the existing exclusion. The exclusion cannot be reinstated before
        the governed cool-off (origination.self_exclusion.min_exclusion_days) has
        elapsed.
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [programme_id, msisdn_token, channel]
              properties:
                programme_id: { type: string }
                msisdn_token: { type: string }
                channel: { type: string, description: Must be a governed allowed_channel }
                reason: { type: string }
      responses:
        "200":
          description: Exclusion recorded (or the existing one, if already excluded)
          content:
            application/json:
              schema:
                type: object
                properties:
                  exclusion_id: { type: string }
                  min_until: { type: string, format: date-time, description: Earliest reinstatement (cool-off end) }
                  already_excluded: { type: boolean }
        "400":
          description: Malformed request or ungoverned channel (SELF_EXCLUSION_BAD_REQUEST / CHANNEL_NOT_ALLOWED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/self-exclusions/reinstate:
    post:
      operationId: reinstateSelfExclusion
      summary: Reinstate a self-excluded subscriber after the cool-off
      description: >
        Lifts a self-exclusion once the governed cool-off has elapsed. A
        reinstatement before the cool-off is refused (COOL_OFF_NOT_ELAPSED).
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [programme_id, msisdn_token, channel]
              properties:
                programme_id: { type: string }
                msisdn_token: { type: string }
                channel: { type: string }
      responses:
        "200":
          description: Subscriber reinstated
          content:
            application/json:
              schema:
                type: object
                properties:
                  msisdn_token: { type: string }
                  state: { type: string, enum: [REINSTATED] }
        "409":
          description: No active exclusion, cool-off not elapsed, or operator reinstatement required (NOT_SELF_EXCLUDED / COOL_OFF_NOT_ELAPSED / OPERATOR_REINSTATEMENT_REQUIRED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Telco/channel credential. Only the SHA-256 of the key is stored
        (V2-SEC-005). The credential — not any payload field — determines the
        tenant context.
    # AdminKeyAuth (X-Admin-Key) removed in EXT-1 — the header-authenticated
    # admin config API it secured was a role-unaware bypass of the portal's
    # RBAC + scope model. Operator auth is the session cookie only.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        Required on all mutating commands (V2-API-002). Scoped to
        (tenant, operation). A retry with the same key returns the original
        outcome; it never repeats the economic effect.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        BC-6 lineage id. Accepted when 1..64 chars of [A-Za-z0-9_.-];
        otherwise discarded and re-minted (never truncated). Always echoed on
        the response and carried through every command, event, journal and
        audit row this request causes.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9_.-]+$"
  schemas:
    Money:
      type: object
      description: Integer minor units + ISO 4217 currency (ADR-0002; never floats).
      required: [amount_minor, currency]
      properties:
        amount_minor: { type: integer, format: int64 }
        currency: { type: string, minLength: 3, maxLength: 3, example: NGN }
    Offer:
      type: object
      required: [offer_id, face_value, fee, disbursed, repayment, fee_model, expires_at, disclosure_ref, disclosure_text, total_cost_text]
      properties:
        offer_id: { type: string }
        face_value: { $ref: "#/components/schemas/Money" }
        fee: { $ref: "#/components/schemas/Money" }
        disbursed: { $ref: "#/components/schemas/Money" }
        repayment: { $ref: "#/components/schemas/Money" }
        fee_model: { type: string, enum: [DEDUCTED_UPFRONT, ADDED_TO_REPAYMENT] }
        expires_at: { type: string, format: date-time }
        disclosure_ref:
          type: string
          description: >
            R-P0-7: reference to the disclosure snapshot presented for this
            offer. The channel MUST render disclosure_text and echo this ref
            back at confirm as proof of what was shown.
        disclosure_text: { type: string, description: The exact disclosure to render to the customer. }
        total_cost_text: { type: string, description: The total-cost representation (what the customer will repay). }
        disclosure_locale: { type: string }
        template_id: { type: string }
        template_version: { type: string }
        disclosure_expires_at: { type: string, format: date-time }
    Advance:
      type: object
      description: Customer-safe advance view (V2-ADV-016 — internal ambiguity never exposed).
      required: [advance_id, status, face_value, disbursed, outstanding, status_route]
      properties:
        advance_id: { type: string }
        status:
          type: string
          enum: [PROCESSING, ACTIVE, PARTIALLY_RECOVERED, CLOSED, FAILED]
        face_value: { $ref: "#/components/schemas/Money" }
        disbursed: { $ref: "#/components/schemas/Money" }
        outstanding: { $ref: "#/components/schemas/Money" }
        status_route: { type: string, example: /v1/advances/adv_01J... }
...
    Error:
      type: object
      description: Standard error envelope (V2-API-011).
      required: [error_code, message, retryable]
      properties:
        error_code:
          type: string
          description: Stable business error code, family-prefixed.
          example: TENANT_CONTEXT_MISMATCH
        message:
          type: string
          description: Safe human-readable message; never stack traces or secrets.
        retryable:
          type: boolean
        correlation_id:
          type: string
  responses:
    Unauthorized:
      description: Missing, unknown or revoked credential (AUTH_INVALID_CLIENT).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...