# backend/internal/handler/channel.go
package handler

// Channel + recovery HTTP surface (M1b-3 tail). BC-7: typed domain errors map
// to stable error-code families exactly once, here. BC-6: the correlation id
// is minted/accepted at this edge and rides every downstream artifact.
// V2-ADV-016: internal ambiguity is never exposed — customers see a safe
// status with a durable status route.

import (
	"context"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"time"

	"github.com/ArowuTest/telco-credit-platform/backend/internal/entity"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/platform"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/platform/ratelimit"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/repo"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/usecase/origination"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/usecase/recovery"
	"github.com/ArowuTest/telco-credit-platform/backend/internal/usecase/treasury"
)

const headerCorrelationID = "X-Correlation-Id"
const headerIdempotencyKey = "Idempotency-Key"

// Correlation wraps a handler with BC-6 correlation propagation: accept the
// caller's id or mint one; always echo it on the response.
//
// VR10-F1: the caller-supplied id is BOUNDED (<=64 chars, [A-Za-z0-9_.-])
// before it can persist into immutable journal rows and logs — anything
// outside the envelope is discarded and re-minted, never truncated (a
// truncated id is a lineage lie).
func Correlation(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		cor := r.Header.Get(headerCorrelationID)
		if !validCorrelationID(cor) {
			cor = platform.NewID("cor")
		}
		w.Header().Set(headerCorrelationID, cor)
		next.ServeHTTP(w, r.WithContext(platform.WithCorrelation(r.Context(), cor)))
	})
}

func validCorrelationID(s string) bool {
	if len(s) == 0 || len(s) > 64 {
		return false
	}
	for _, r := range s {
		switch {
		case r >= 'a' && r <= 'z', r >= 'A' && r <= 'Z', r >= '0' && r <= '9',
			r == '_', r == '-', r == '.':
		default:
			return false
		}
	}
	return true
}

// Channel serves the customer-facing credit journey.
type Channel struct {
	Origination       *origination.Service
	Recovery          *recovery.Service
	Limiter           *ratelimit.Limiter // R-P0-8 inbound rate limit (channel API)
	TrustedProxyCount int                // R-P2-7 client-IP derivation
	Log               *slog.Logger
}

// Mount registers routes; auth resolves the tenant, Correlation the lineage.
// R-P0-8 (+F1): a PRE-auth IP throttle (channel_ip) fires before any DB
// lookup, so a rotating-invalid-key flood — which never resolves a telco — is
// stopped by its shared IP bucket; the per-telco fairness throttle then runs
// POST-auth on the validated telco.
func (h *Channel) Mount(mux *http.ServeMux, auth *TenantAuth) {
	if h.Limiter == nil {
		panic("channel: rate limiter is required (R-P0-8 fail-closed)")
	}
	ipKey := func(r *http.Request) string { return clientIP(r, h.TrustedProxyCount) }
	wrap := func(fn http.HandlerFunc) http.Handler {
		return rateLimited(h.Limiter, "channel_ip", ipKey,
			auth.Wrap(perTelcoRateLimit(h.Limiter, Correlation(fn))))
	}
	// BX-MED-007: offers is a read, but it is keyed by the subscriber msisdn_token — a sensitive
	// identifier that must never ride in a URL (query strings leak into browser history, access
	// logs, proxies and telemetry). It is a POST so the token travels in the request body.
	mux.Handle("POST /v1/offers", wrap(h.getOffers))
	mux.Handle("POST /v1/advances", wrap(h.confirm))
	mux.Handle("GET /v1/advances/{id}", wrap(h.advanceStatus))
	// Recovery events have exactly ONE external ingress: the HMAC-signed recharge
	// webhook (POST /v1/telcos/{telco}/recharge-webhook), which namespaces every
	// event "wh:" so EOD reconciliation covers it (recon selects LIKE 'wh:%') and
	// routes it through the per-event/daily HELD blast-radius clamps. A general
	// api-key channel route here would book money INVISIBLE to recon and skip the
	// clamps + the S3-C2 re-origination hold — do NOT re-add one (BX-P0-002).
	// Self-exclusion (R1-MUST): a subscriber opts out of credit; reinstatement is
	// governed by a cool-off enforced in the usecase.
	mux.Handle("POST /v1/self-exclusions", wrap(h.selfExclude))
	mux.Handle("POST /v1/self-exclusions/reinstate", wrap(h.reinstateSelfExclusion))
}

// --- wire shapes -----------------------------------------------------------

type offerResponse struct {
	OfferID   string       `json:"offer_id"`
	FaceValue entity.Money `json:"face_value"`
	Fee       entity.Money `json:"fee"`
	Disbursed entity.Money `json:"disbursed"`
	Repayment entity.Money `json:"repayment"`
	FeeModel  string       `json:"fee_model"`
	ExpiresAt time.Time    `json:"expires_at"`
	// R-P0-7: the disclosure the channel MUST render, and the reference to echo
	// back at confirm as proof of what was shown.
	DisclosureRef     string    `json:"disclosure_ref"`
	DisclosureText    string    `json:"disclosure_text"`
	TotalCostText     string    `json:"total_cost_text"`
	DisclosureLocale  string    `json:"disclosure_locale"`
	TemplateID        string    `json:"template_id"`
	TemplateVersion   string    `json:"template_version"`
	DisclosureExpires time.Time `json:"disclosure_expires_at"`
}

type confirmRequest struct {
	ProgrammeID string `json:"programme_id"`
	OfferID     string `json:"offer_id"`
	MSISDNToken string `json:"msisdn_token"`
	// R-P0-7 consent/channel evidence. Channel/SessionID/AcceptedAt are
	// telco-supplied (DD-06); DisclosureRef echoes the presented disclosure;
	// TelcoEvidence is the optional telco acceptance signature.
	DisclosureRef string          `json:"disclosure_ref"`
	Channel       string          `json:"channel"`
	SessionID     string          `json:"session_id"`
	AcceptedAt    time.Time       `json:"accepted_at"`
	TelcoEvidence json.RawMessage `json:"telco_evidence,omitempty"`
}

type advanceResponse struct {
	AdvanceID   string       `json:"advance_id"`
	Status      string       `json:"status"` // customer-safe (V2-ADV-016)
	FaceValue   entity.Money `json:"face_value"`
	Disbursed   entity.Money `json:"disbursed"`
	Outstanding entity.Money `json:"outstanding"`
	StatusRoute string       `json:"status_route"` // EDG-004: durable status URL
}

// customerStatus maps internal FSM states to the customer-safe vocabulary
// (V2-ADV-016): ambiguity and pre-fulfilment states read as PROCESSING.
func customerStatus(s entity.AdvanceState) string {
	switch s {
	case entity.AdvRequested, entity.AdvValidated, entity.AdvExposureReserved,
		entity.AdvPendingFulfilment, entity.AdvFulfilmentUnknown:
		return "PROCESSING"
	case entity.AdvFulfilmentFailed, entity.AdvDeclined:
		return "FAILED"
	default:
		return string(s) // ACTIVE, PARTIALLY_RECOVERED, CLOSED
	}
}

func toAdvanceResponse(a entity.Advance) advanceResponse {
	return advanceResponse{
		AdvanceID:   a.AdvanceID,
		Status:      customerStatus(a.State),
		FaceValue:   a.FaceValue,
		Disbursed:   a.Disbursed,
		Outstanding: a.Outstanding,
		StatusRoute: "/v1/advances/" + a.AdvanceID,
	}
}

// --- handlers --------------------------------------------------------------

type offersRequest struct {
	ProgrammeID string `json:"programme_id"`
	MSISDNToken string `json:"msisdn_token"`
}

func (h *Channel) getOffers(w http.ResponseWriter, r *http.Request) {
	// BX-MED-007: the msisdn_token comes from the POST body, never a query string.
	var req offersRequest
	if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, 1<<16)).Decode(&req); err != nil {
		writeErr(w, http.StatusBadRequest, "OFFER_BAD_REQUEST", "malformed JSON body")
		return
	}
	if req.ProgrammeID == "" || req.MSISDNToken == "" {
		writeErr(w, http.StatusBadRequest, "OFFER_BAD_REQUEST", "programme_id and msisdn_token are required")
		return
	}
	offers, err := h.Origination.GetOffers(r.Context(), req.ProgrammeID, req.MSISDNToken)
	if err != nil {
		h.writeDomainErr(w, r, err)
		return
	}
	out := make([]offerResponse, 0, len(offers))
	for _, ov := range offers {
		o, d := ov.Offer, ov.Disclosure
		out = append(out, offerResponse{
			OfferID: o.OfferID, FaceValue: o.FaceValue, Fee: o.Fee,
			Disbursed: o.Disbursed, Repayment: o.Repayment,
			FeeModel: string(o.FeeModel), ExpiresAt: o.ExpiresAt,
			DisclosureRef: d.DisclosureSnapshotID, DisclosureText: d.RenderedBody,
			TotalCostText: d.TotalCostText, DisclosureLocale: d.Locale,
			TemplateID: d.TemplateID, TemplateVersion: d.TemplateVersion,
			DisclosureExpires: d.ExpiresAt,
		})
	}
	writeJSON(w, http.StatusOK, out)
}

func (h *Channel) confirm(w http.ResponseWriter, r *http.Request) {
	idemKey := r.Header.Get(headerIdempotencyKey)
	if len(idemKey) < 8 || len(idemKey) > 128 {
		// V2-API-002: mutating commands REQUIRE a usable idempotency key.
		writeErr(w, http.StatusBadRequest, "ADVANCE_IDEMPOTENCY_KEY_REQUIRED",
			"Idempotency-Key header (8..128 chars) is required")
		return
	}
	var req confirmRequest
	if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, 1<<16)).Decode(&req); err != nil {
		writeErr(w, http.StatusBadRequest, "ADVANCE_BAD_REQUEST", "malformed JSON body")
		return
	}
	res, err := h.Origination.Confirm(r.Context(), origination.ConfirmCmd{
		ProgrammeID:   req.ProgrammeID,
		OfferID:       req.OfferID,
		MSISDNToken:   req.MSISDNToken,
		IdemKey:       idemKey,
		CorrelationID: platform.CorrelationFrom(r.Context()),
		DisclosureRef: req.DisclosureRef,
		Channel:       req.Channel,
		SessionID:     req.SessionID,
		AcceptedAt:    req.AcceptedAt,
		TelcoEvidence: req.TelcoEvidence,
	})
	if err != nil {
		h.writeDomainErr(w, r, err)
		return
	}
	status := http.StatusCreated
	switch {
	case res.Replayed:
		status = http.StatusOK // EDG-001: replay returns the original outcome
	case res.Advance.State == entity.AdvFulfilmentUnknown:
		status = http.StatusAccepted // EDG-004: safe pending + status route
	case res.Advance.State == entity.AdvFulfilmentFailed:
		status = http.StatusUnprocessableEntity
	}
	writeJSON(w, status, toAdvanceResponse(res.Advance))
}

func (h *Channel) advanceStatus(w http.ResponseWriter, r *http.Request) {
	adv, err := h.Origination.GetAdvance(r.Context(), r.PathValue("id"))
	if errors.Is(err, repo.ErrNotFound) {
		// VR10-F2: this route's 404 is about the ADVANCE, not an offer.
		writeErr(w, http.StatusNotFound, "ADVANCE_NOT_FOUND", "advance not found")
		return
	}
	if err != nil {
		h.writeDomainErr(w, r, err)
		return
	}
	writeJSON(w, http.StatusOK, toAdvanceResponse(adv))
}

// writeDomainErr is THE typed-error boundary (BC-7): every domain error family
// renders its documented envelope; anything unmapped is a 500 with no detail
// leaked (V2-API-011).
func (h *Channel) writeDomainErr(w http.ResponseWriter, r *http.Request, err error) {
	switch {
	case errors.Is(err, origination.ErrOfferNotFound), errors.Is(err, repo.ErrNotFound):
		writeErr(w, http.StatusNotFound, "OFFER_NOT_FOUND", "offer or resource not found")
	case errors.Is(err, origination.ErrOfferExpired):
		writeErr(w, http.StatusConflict, "OFFER_EXPIRED", "offer expired; request fresh offers")
	case errors.Is(err, origination.ErrOfferNotAcceptable):
		writeErr(w, http.StatusConflict, "OFFER_SNAPSHOT_MISMATCH", "offer no longer acceptable; request fresh offers")
	case errors.Is(err, origination.ErrProgrammeMismatch):
		// BX-P0-001: the confirm named a programme other than the offer's — refuse so
		// no gate (kill-switch/economics/disclosure) can be evaluated against a
		// different programme than the money is booked under.
		writeErr(w, http.StatusConflict, "PROGRAMME_MISMATCH", "the offer belongs to a different programme; request fresh offers")
	case errors.Is(err, origination.ErrDivergentDuplicate):
		// R-P0-1: same idempotency key, different request. Not a replay —
		// the client must use a fresh key for a new command.
		writeErr(w, http.StatusConflict, "DIVERGENT_DUPLICATE", "idempotency key already used for a different request; use a fresh key")
	case errors.Is(err, recovery.ErrDivergentRecovery):
		// R-P0-2: same source_event_id, different payload — refused, audited.
		writeErr(w, http.StatusConflict, "DIVERGENT_DUPLICATE", "source event id already used for a different payload")
	case errors.Is(err, origination.ErrSubscriberIneligible):
		writeErr(w, http.StatusForbidden, "SUBSCRIBER_INACTIVE", "subscriber not eligible")
	case errors.Is(err, origination.ErrRecoveryUnconfirmedHold):
		// S3-C2: a debt closed by a not-yet-confirmed webhook recovery holds
		// re-origination until the EOD recon confirms it (fail-closed).
		writeErr(w, http.StatusConflict, "RECOVERY_UNCONFIRMED_HOLD", "a recent recovery is pending confirmation; re-borrowing is temporarily on hold")
	case errors.Is(err, repo.ErrConcurrentAdvanceBlocked):
		writeErr(w, http.StatusConflict, "CONCURRENT_ADVANCE_BLOCK", "an advance is already open for this subscriber")
	case errors.Is(err, repo.ErrOfferAlreadyUsed):
		writeErr(w, http.StatusConflict, "OFFER_SNAPSHOT_MISMATCH", "offer already used")
	case errors.Is(err, repo.ErrNoFundingCapacity):
		// V2 §6.2 FUNDING_*: retry only after treasury state changes.
		writeErr(w, http.StatusConflict, "FUNDING_POOL_EXHAUSTED", "no funding capacity available")
	case errors.Is(err, origination.ErrDecisionUnavailable):
		// EDG-014 boundary: stale/absent decision is a customer-safe no-offer.
		writeErr(w, http.StatusConflict, "NO_OFFER_AVAILABLE", "no offer available at this time")
	case errors.Is(err, origination.ErrDisclosureRequired),
		errors.Is(err, origination.ErrAcceptanceEvidenceMissing):
		// R-P0-7: a confirm without disclosure/channel/session/acceptance
		// evidence is malformed — the client must re-fetch offers and accept.
		writeErr(w, http.StatusBadRequest, "DISCLOSURE_EVIDENCE_REQUIRED", "disclosure reference and channel/session/acceptance evidence are required")
	case errors.Is(err, origination.ErrDisclosureMismatch),
		errors.Is(err, origination.ErrDisclosureExpired):
		// R-P0-7: the echoed disclosure does not match the offer, or acceptance
		// fell outside its validity — request fresh offers and re-disclose.
		writeErr(w, http.StatusConflict, "DISCLOSURE_MISMATCH", "disclosure no longer valid for this offer; request fresh offers")
	case errors.Is(err, origination.ErrChannelNotAllowed):
		writeErr(w, http.StatusBadRequest, "CHANNEL_NOT_ALLOWED", "this channel is not permitted for this programme")
	case errors.Is(err, origination.ErrSelfExclusionChannelNotAllowed):
		writeErr(w, http.StatusBadRequest, "CHANNEL_NOT_ALLOWED", "this channel is not permitted for self-exclusion")
	case errors.Is(err, origination.ErrNotSelfExcluded):
		writeErr(w, http.StatusConflict, "NOT_SELF_EXCLUDED", "no active self-exclusion to reinstate")
	case errors.Is(err, origination.ErrCoolOffNotElapsed):
		writeErr(w, http.StatusConflict, "COOL_OFF_NOT_ELAPSED", "the self-exclusion cool-off period has not elapsed")
	case errors.Is(err, origination.ErrOperatorReinstatementRequired):
		writeErr(w, http.StatusConflict, "OPERATOR_REINSTATEMENT_REQUIRED", "reinstatement requires operator approval; self-service is not available")
	case errors.Is(err, origination.ErrDisclosureUnavailable):
		// Fail-closed: no active disclosure policy — cannot disclose, cannot serve.
		writeErr(w, http.StatusServiceUnavailable, "SERVICE_TEMPORARILY_LIMITED", "service temporarily limited; try again later")
	case errors.Is(err, origination.ErrOverlayBlocked):
		// V2-SCR-015: which flag fired is logged upstream, never disclosed.
		writeErr(w, http.StatusForbidden, "SERVICE_RESTRICTED", "service not available for this subscriber right now")
	case errors.Is(err, origination.ErrProgrammeEconomicsNotSet), errors.Is(err, origination.ErrProgrammeEconomicsInvalid):
		// Build 1 fail-closed go-live gate: the programme cannot originate until its
		// economic/legal identity is configured. Non-revealing to the customer; ops
		// sees the specific error in logs.
		writeErr(w, http.StatusServiceUnavailable, "SERVICE_TEMPORARILY_LIMITED", "service temporarily limited; try again later")
	case errors.Is(err, treasury.ErrProgrammeSuspended):
		// M3d fail-closed: lending stopped (guardrail trip or operator).
		writeErr(w, http.StatusServiceUnavailable, "SERVICE_TEMPORARILY_LIMITED", "service temporarily limited; try again later")
	case errors.Is(err, context.DeadlineExceeded):
		writeErr(w, http.StatusGatewayTimeout, "SYSTEM_TEMPORARILY_UNAVAILABLE", "timeout")
	default:
		h.Log.Error("unmapped domain error", "err", err, "path", r.URL.Path)
		writeErr(w, http.StatusInternalServerError, "SYSTEM_TEMPORARILY_UNAVAILABLE", "internal error")
	}
}

// --- self-exclusion (R1-MUST) ----------------------------------------------

type selfExclusionRequest struct {
	ProgrammeID string `json:"programme_id"`
	MSISDNToken string `json:"msisdn_token"`
	Channel     string `json:"channel"`
	Reason      string `json:"reason"`
}

type selfExclusionResponse struct {
	ExclusionID     string    `json:"exclusion_id"`
	MinUntil        time.Time `json:"min_until"`
	AlreadyExcluded bool      `json:"already_excluded"`
}

// selfExclude records a subscriber's opt-out from credit. The tenant is resolved
// by auth into the context; enforcement is at the offer/confirm gate.
func (h *Channel) selfExclude(w http.ResponseWriter, r *http.Request) {
	var req selfExclusionRequest
	if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, 1<<14)).Decode(&req); err != nil {
		writeErr(w, http.StatusBadRequest, "SELF_EXCLUSION_BAD_REQUEST", "malformed JSON body")
		return
	}
	if req.ProgrammeID == "" || req.MSISDNToken == "" || req.Channel == "" {
		writeErr(w, http.StatusBadRequest, "SELF_EXCLUSION_BAD_REQUEST", "programme_id, msisdn_token and channel are required")
		return
	}
	res, err := h.Origination.RequestSelfExclusion(r.Context(), req.ProgrammeID, req.MSISDNToken, req.Channel, req.Reason)
	if err != nil {
		h.writeDomainErr(w, r, err)
		return
	}
	writeJSON(w, http.StatusOK, selfExclusionResponse{
		ExclusionID: res.ExclusionID, MinUntil: res.MinUntil, AlreadyExcluded: res.AlreadyExcluded,
	})
}

// reinstateSelfExclusion lifts a self-exclusion once the governed cool-off has
// elapsed (the usecase refuses an early reinstatement).
func (h *Channel) reinstateSelfExclusion(w http.ResponseWriter, r *http.Request) {
	var req selfExclusionRequest
	if err := json.NewDecoder(http.MaxBytesReader(w, r.Body, 1<<14)).Decode(&req); err != nil {
		writeErr(w, http.StatusBadRequest, "SELF_EXCLUSION_BAD_REQUEST", "malformed JSON body")
		return
	}
	if req.ProgrammeID == "" || req.MSISDNToken == "" || req.Channel == "" {
		writeErr(w, http.StatusBadRequest, "SELF_EXCLUSION_BAD_REQUEST", "programme_id, msisdn_token and channel are required")
		return
	}
	if err := h.Origination.ReinstateSelfExclusion(r.Context(), req.ProgrammeID, req.MSISDNToken, req.Channel); err != nil {
		h.writeDomainErr(w, r, err)
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"msisdn_token": req.MSISDNToken, "state": "REINSTATED"})
}